LPS-69876 Refactor both DDC into base class
