+++ modules/apps/collaboration/blogs/blogs-demo-data-creator-impl/bnd.bnd
@@ -1,3 +1,4 @@
 Bundle-Name: Liferay Blogs Demo Data Creator Implementation
 Bundle-SymbolicName: com.liferay.blogs.demo.data.creator.impl
 Bundle-Version: 1.0.0
+-dsannotations-options: inherit
