+++ modules/apps/collaboration/blogs/blogs-demo-data-creator-impl/src/main/java/com/liferay/blogs/demo/data/creator/internal/CreativeCommonsBlogsEntryDemoDataCreatorImpl.java
@@ -15,30 +15,14 @@
 package com.liferay.blogs.demo.data.creator.internal;
 
 import com.liferay.blogs.demo.data.creator.BlogsEntryDemoDataCreator;
-import com.liferay.blogs.exception.NoSuchEntryException;
 import com.liferay.blogs.model.BlogsEntry;
-import com.liferay.blogs.service.BlogsEntryLocalService;
-import com.liferay.document.library.demo.data.creator.FileEntryDemoDataCreator;
-import com.liferay.document.library.demo.data.creator.RootFolderDemoDataCreator;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.repository.model.FileVersion;
-import com.liferay.portal.kernel.repository.model.Folder;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.servlet.taglib.ui.ImageSelector;
-import com.liferay.portal.kernel.util.FileUtil;
-import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 
 import java.io.IOException;
 
-import java.sql.Timestamp;
-
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.Date;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -46,14 +30,13 @@ import java.util.concurrent.atomic.AtomicInteger;
 import org.osgi.framework.BundleContext;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Sergio González
  */
 @Component(service = BlogsEntryDemoDataCreator.class)
 public class CreativeCommonsBlogsEntryDemoDataCreatorImpl
-	implements BlogsEntryDemoDataCreator {
+	extends BaseBlogsEntryDemoDataCreator {
 
 	@Activate
 	public void activate(BundleContext bundleContext) {
@@ -66,68 +49,15 @@ public class CreativeCommonsBlogsEntryDemoDataCreatorImpl
 	public BlogsEntry create(long userId, long groupId)
 		throws IOException, PortalException {
 
-		ServiceContext serviceContext = new ServiceContext();
-
-		serviceContext.setAddGroupPermissions(true);
-		serviceContext.setAddGuestPermissions(true);
-
-		serviceContext.setScopeGroupId(groupId);
-
-		ImageSelector imageSelector = new ImageSelector(
-			_getRandomImageBytes(userId, groupId),
-			StringUtil.randomString() + ".jpeg", "image/jpeg",
-			StringPool.BLANK);
-
 		int index = _getNextIndex();
 
-		BlogsEntry blogsEntry = _blogsEntryLocalService.addEntry(
-			userId, _getRandomTitle(index), _getRandomSubtitle(index), null,
-			_getRandomContent(index), _getRandomDate(), false, false, null,
-			null, imageSelector, null, serviceContext);
+		String title = _getRandomTitle(index);
 
-		_entryIds.add(blogsEntry.getEntryId());
+		String subtitle = _getRandomSubtitle(index);
 
-		return blogsEntry;
-	}
+		String content = _getRandomContent(index);
 
-	@Override
-	public void delete() throws PortalException {
-		try {
-			for (long entryId : _entryIds) {
-				_blogsEntryLocalService.deleteEntry(entryId);
-
-				_entryIds.remove(entryId);
-			}
-		}
-		catch (NoSuchEntryException nsee) {
-			if (_log.isWarnEnabled()) {
-				_log.warn(nsee);
-			}
-		}
-
-		_fileEntryDemoDataCreator.delete();
-		_rootFolderDemoDataCreator.delete();
-	}
-
-	@Reference(unbind = "-")
-	protected void setBlogLocalService(
-		BlogsEntryLocalService blogsEntryLocalService) {
-
-		_blogsEntryLocalService = blogsEntryLocalService;
-	}
-
-	@Reference(unbind = "-")
-	protected void setFileEntryDemoDataCreator(
-		FileEntryDemoDataCreator fileEntryDemoDataCreator) {
-
-		_fileEntryDemoDataCreator = fileEntryDemoDataCreator;
-	}
-
-	@Reference(unbind = "-")
-	protected void setRootFolderDemoDataCreator(
-		RootFolderDemoDataCreator rootFolderDemoDataCreator) {
-
-		_rootFolderDemoDataCreator = rootFolderDemoDataCreator;
+		return createBaseBlogsEntry(userId, groupId, title, subtitle, content);
 	}
 
 	private int _getNextIndex() {
@@ -144,46 +74,18 @@ public class CreativeCommonsBlogsEntryDemoDataCreatorImpl
 		Class<?> clazz = getClass();
 
 		String titlePath =
-			"com/liferay/blogs/demo/data/creator/internal/dependencies/" +
-				"creative/commons/content" + index + ".txt";
+			"com/liferay/blogs/demo/data/creator/internal/dependencies" +
+				"/creative/commons/content" + index + ".txt";
 
 		return StringUtil.read(clazz.getClassLoader(), titlePath, false);
-	}
-
-	private Date _getRandomDate() {
-		long start = Timestamp.valueOf("2000-01-01 00:00:00").getTime();
-		long end = new Date().getTime();
-
-		long diff = end - start + 1;
-
-		Timestamp timestamp = new Timestamp(
-			start + (long)(Math.random() * diff));
-
-		return new Date(timestamp.getTime());
-	}
-
-	private byte[] _getRandomImageBytes(long userId, long groupId)
-		throws IOException, PortalException {
-
-		if (_blogsEntryImagesFolder == null) {
-			_blogsEntryImagesFolder = _rootFolderDemoDataCreator.create(
-				userId, groupId, "Blogs Images");
-		}
-
-		FileEntry fileEntry = _fileEntryDemoDataCreator.create(
-			userId, _blogsEntryImagesFolder.getFolderId());
-
-		FileVersion fileVersion = fileEntry.getFileVersion();
-
-		return FileUtil.getBytes(fileVersion.getContentStream(false));
 	}
 
 	private String _getRandomSubtitle(int index) throws IOException {
 		Class<?> clazz = getClass();
 
 		String titlePath =
-			"com/liferay/blogs/demo/data/creator/internal/dependencies/" +
-				"creative/commons/subtitle" + index + ".txt";
+			"com/liferay/blogs/demo/data/creator/internal/dependencies" +
+				"/creative/commons/subtitle" + index + ".txt";
 
 		return StringUtil.read(clazz.getClassLoader(), titlePath, false);
 	}
@@ -192,22 +94,14 @@ public class CreativeCommonsBlogsEntryDemoDataCreatorImpl
 		Class<?> clazz = getClass();
 
 		String titlePath =
-			"com/liferay/blogs/demo/data/creator/internal/dependencies/" +
-				"creative/commons/title" + index + ".txt";
+			"com/liferay/blogs/demo/data/creator/internal/dependencies" +
+				"/creative/commons/title" + index + ".txt";
 
 		return StringUtil.read(clazz.getClassLoader(), titlePath, false);
 	}
 
-	private static final Log _log = LogFactoryUtil.getLog(
-		CreativeCommonsBlogsEntryDemoDataCreatorImpl.class);
-
 	private final AtomicInteger _atomicInteger = new AtomicInteger(0);
 	private final List<Integer> _availableIndexes =
 		new CopyOnWriteArrayList<>();
-	private Folder _blogsEntryImagesFolder;
-	private BlogsEntryLocalService _blogsEntryLocalService;
-	private final List<Long> _entryIds = new CopyOnWriteArrayList<>();
-	private FileEntryDemoDataCreator _fileEntryDemoDataCreator;
-	private RootFolderDemoDataCreator _rootFolderDemoDataCreator;
 
 }
